Initial Commit for Software Engineering Methods Coursework
This is the basic outline for the program. It has users classes for
Cook, Eater and Administrator. There is also singleton logger file that
saves and loads serialized lists for the users. There are few windows
for basic interactions.
Part of the code for account creation is written. Currently working on
refactoring and reading back saved lists